+++ CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 
 This changelog was automatically generated using [Caretaker](https://github.com/DevelopersToolbox/caretaker) by [Wolf Software](https://github.com/WolfSoftware)
 
-### [Unreleased](https://github.com/DockerToolbox/goenv/compare/v0.1.0...HEAD)
+### [v0.1.1](https://github.com/DockerToolbox/goenv/compare/v0.1.0...v0.1.1)
 
-- Update pinned packages [`[head]`](https://github.com/DockerToolbox/goenv/commit/)
+> Released on May, 24th 2021
+
+- Added a new 'generate' option to auto-generate (update) the Dockerfiles [`[head]`](https://github.com/DockerToolbox/goenv/commit/)
+
+- Update pinned packages [`[c949f22]`](https://github.com/DockerToolbox/goenv/commit/c949f22ab527350cf91033e64325c11b1416b8c5)
 
 - Fix travis rvm versions and slack integration [`[a987387]`](https://github.com/DockerToolbox/goenv/commit/a987387d86fe6c13b9311a3fba9f8a235234e6fd)
+
+### [v0.1.0](https://github.com/DockerToolbox/goenv/releases/v0.1.0)
+
+> Released on February, 19th 2021
 
 - The initial commit [`[63bdd44]`](https://github.com/DockerToolbox/goenv/commit/63bdd447ef11b36777bb0f395024ddb8cf593625)